Changes albumRating (float -> int) + drops photoID

+++ DTOs/Album/UpdateAlbumDTO.cs
@@ -11,7 +11,7 @@ namespace AlbumAPI.DTOs.Album
         public string YearReleased { get; set; } = string.Empty;
         public string AlbumGenre { get; set; } = string.Empty;
         public string AlbumDescription { get; set; } = string.Empty;
-        public int AlbumRating { get; set; } 
+        public double AlbumRating { get; set; } 
         public string photoURL { get; set; } = string.Empty;
         public string publicID { get; set; } = string.Empty;
     }
